feat(state): reduced complexity, restructure typings, add overloads fro selectSlices
