+++ libs/state/spec/rx-state.component.spec.ts
@@ -8,7 +8,7 @@ import {
   PrimitiveState
 } from './fixtures';
 import { Observable, Subject } from 'rxjs';
-import { select } from '../src/lib/core/operators/select';
+import { select } from '../src/lib/rxjs/operators/select';
 
 const initialParentState: PrimitiveState = {
   ...initialPrimitiveState,
