+++ libs/state/spec/rxjs/operators/stateful.spec.ts
@@ -2,7 +2,7 @@ import { jestMatcher } from '@test-helpers';
 import { EMPTY, NEVER } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { TestScheduler } from 'rxjs/testing';
-import { stateful } from '../../../src';
+import { stateful } from '@rx-angular/state';
 import { initialNestedState } from '../../fixtures';
 
 let testScheduler: TestScheduler;
