+++ libs/state/src/lib/core/index.ts
@@ -1,4 +1,4 @@
 export * from './utils';
-export * from './operators';
+export * from '../rxjs/operators';
 export * from './accumulation-observable';
 export * from './side-effect-observable';
