+++ libs/state/src/lib/index.ts
@@ -1,2 +1,3 @@
 export { RxState } from './rx-state.service';
-export { select, stateful, distinctUntilSomeChanged } from './core/operators';
+export { select, stateful, distinctUntilSomeChanged } from './rxjs/operators';
+export { KeyCompareMap, CompareFn } from './rxjs/interfaces';
