+++ libs/state/src/lib/rxjs/interfaces/index.ts
@@ -1,0 +1,2 @@
+export { CompareFn } from './compare-fn';
+export { KeyCompareMap } from './key-compare-map';
