+++ libs/state/src/lib/rxjs/interfaces/key-compare-map.ts
@@ -1,0 +1,23 @@
+/**
+ * @description
+ * The `KeyCompareMap` is used to configure custom comparison for defined keys. You can set the `CompareFn` to
+ * `undefined` in order to utilize the default equality check.
+ *
+ * @example
+ * const keyCompareMap = {
+ *    myKey: (o, n) => customCompare(o, n),
+ *    myOtherKey: undefined
+ *  };
+ *  const o$ = of({
+ *    myKey: 5,
+ *    myOtherKey: 'bar'
+ *  }).pipe(distinctUntilSomeChanged(keyCompareMap));
+ *
+ * @docsPage distinctUntilSomeChanged
+ * @docsCategory operators
+ */
+import { CompareFn } from './compare-fn';
+
+export type KeyCompareMap<T extends object> = {
+  [K in keyof T]?: CompareFn<T[K]>;
+};
